[TM-1742] Rename to TreeSpeciesTable.
