+++ src/pages/nursery/[uuid]/tabs/Overview.tsx
@@ -6,7 +6,7 @@ import PageBody from "@/components/extensive/PageElements/Body/PageBody";
 import PageCard from "@/components/extensive/PageElements/Card/PageCard";
 import PageColumn from "@/components/extensive/PageElements/Column/PageColumn";
 import PageRow from "@/components/extensive/PageElements/Row/PageRow";
-import TreeSpeciesTablePD from "@/components/extensive/Tables/TreeSpeciesTablePD";
+import TreeSpeciesTable from "@/components/extensive/Tables/TreeSpeciesTable";
 import { useDate } from "@/hooks/useDate";
 
 interface NurseryOverviewTabProps {
@@ -36,7 +36,7 @@ const NurseryOverviewTab = ({ nursery }: NurseryOverviewTabProps) => {
             <TextField label={t("Seedlings or Young Trees to be Grown")} value={nursery?.seedling_grown} />
           </PageCard>
           <PageCard title={"Tree Species"}>
-            <TreeSpeciesTablePD entityUuid={nursery?.uuid} entity="nurseries" collection="nursery-seedling" />
+            <TreeSpeciesTable entityUuid={nursery?.uuid} entity="nurseries" collection="nursery-seedling" />
           </PageCard>
         </PageColumn>
       </PageRow>
